requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python
pillow
numpy
transformers
torch
torchvision
